tests/wpt/tests/html/canvas/offscreen/layers/2d.layer.exceptions-are-no-op.html

<!DOCTYPE html>
<!-- DO NOT EDIT! This test has been generated by /html/canvas/tools/gentest.py. -->
<title>OffscreenCanvas test: 2d.layer.exceptions-are-no-op</title>
<script src="/resources/testharness.js"></script>
<script src="/resources/testharnessreport.js"></script>
<script src="/html/canvas/resources/canvas-tests.js"></script>

<h1>2d.layer.exceptions-are-no-op</h1>
<p class="desc">Checks that the context state is left unchanged if beginLayer throws.</p>


<script>
var t = async_test("Checks that the context state is left unchanged if beginLayer throws.");
var t_pass = t.done.bind(t);
var t_fail = t.step_func(function(reason) {
    throw reason;
});
t.step(function() {

  var canvas = new OffscreenCanvas(100, 50);
  var ctx = canvas.getContext('2d');

  // Get `beginLayer` to throw while parsing the filter.
  assert_throws_js(TypeError,
                   () => ctx.beginLayer({filter: {name: 'colorMatrix',
                                                  values: 'foo'}}));
  // `beginLayer` shouldn't have opened the layer, so `endLayer` should throw.
  assert_throws_dom("InvalidStateError", () => ctx.endLayer());
  t.done();

});
</script>
